requirements.txt
boto3
botocore
click
cloudpickle
dask
datashader
debugpy
dgl
elasticsearch_dsl
fuzzymatcher
geopandas
geopy
hdbscan
holoviews
interchange
joblib
matplotlib
mlflow
neo4j
nest_asyncio
nltk
numpy
ogb
optuna
packaging
pandas
paramiko
plotly
pmdarima
psutil
py2neo
pyarrow
python_dateutil
python_dotenv
pytorch_lightning
pytz
PyYAML
ray
requests
requests_futures
scikit_learn
scipy
scp
smart_open
statsmodels
thefuzz
torch
tqdm
transformers
umap_learn
urllib3
python-dotenv
umap_learn
vespid